Move Automata operations into struct, and move unit tests to corresponding source files
